+++ packages/terra-clinical-action-header/CHANGELOG.md
@@ -4,7 +4,7 @@ ChangeLog
 Unreleased
 ----------
 ### Added
-* prop to keep the close button in small viewports
+* prop `keepCloseButton` to keep the close button in small viewports
 
 1.6.0 - (December 9, 2017)
 ----------